+++ src/client.h
@@ -111,7 +111,7 @@ public:
 			IrrlichtDevice *device,
 			const char *playername,
 			JMutex &range_mutex,
-			s16 &viewing_range_nodes,
+			float &viewing_range_nodes,
 			bool &viewing_range_all
 			);
